Add ARM Coprocessor accessor macros and vector
Arm coprocessor registers are accessed via mrc/mcr instructions such as:

  mrc p15, 0, r0, c0, c0, 0

Where:
        mrc -> Move to arm register from coprocessor
        mcr -> Move to coprocessor from arm register
        r0  -> Source/Destination arm register

        p15/0/c0/c0/0 -> Coprocessor register address.

CP registers are addressed via the form cp/op1/crn/crm/op2.

The fact that the cp register address cannot be in a register makes
putting this into C-friendly form difficult, so the solution is to use a
CPP macro that injects some inline assembler to emit the instruction.

Architected CP registers have macro names defined to make source code
more readable, ex, the following are the same:
    cp_read(myvar, 15, 0, 0, 0, 0);
    cp_read(myvar, MIDR);
Where 'myvar' is the C variable to write the value to (must be a valid
lval).

Change-Id: I601f546079f53f93b9f3af03cfd8e401699e8702
Signed-off-by: Tim Detwiler <timdetwiler@gmail.com>