nios2: Add srec target.
In order to utilize the Altera board flashing tools, we need to get the
kernel into the 'srec' format. This target depends on tools in the
altera toolchain.

Some parameters in this target (base, end, entry) will be board
dependant and are currently filled with place-holder values (meaning the
target is non-functional at this time).

Change-Id: I0b360786622970808a623f452c1eba8e38316075
Signed-off-by: Tim Detwiler <timdetwiler@gmail.com>